Gateway: Create package schemas and routes
